ollama-ocr-demo/requirements.txt
ollama==0.5.1
pillow==11.3.0